+++ com.github.eclipsecolortheme/src/com/github/eclipsecolortheme/mapper/WebEditorMapper.java
@@ -86,7 +86,7 @@ public class WebEditorMapper extends ThemePreferenceMapper {
             putPreference("MEDIA", entry(theme.get(KEYWORD)));
             break;
         case JSP:
-        	putPreference("tagBorder", entry(theme.get(METHOD)));
+            putPreference("tagBorder", entry(theme.get(METHOD)));
             putPreference("tagName", entry(theme.get(METHOD)));
             putPreference("tagAttributeName", entry(theme.get(KEYWORD)));
             putPreference("tagAttributeEquals",
@@ -176,15 +176,15 @@ public class WebEditorMapper extends ThemePreferenceMapper {
             preferences.remove("MEDIA");
             break;
         case JSP:
-        	preferences.remove("tagBorder");
-        	preferences.remove("tagName");
-        	preferences.remove("tagAttributeName");
-        	preferences.remove("tagAttributeEquals");
-        	preferences.remove("tagAttributeValue");
-        	preferences.remove("commentBorder");
-        	preferences.remove("commentText");
-        	preferences.remove("SCRIPT_AREA_BORDER");
-        	preferences.remove("jsp_content");
+            preferences.remove("tagBorder");
+            preferences.remove("tagName");
+            preferences.remove("tagAttributeName");
+            preferences.remove("tagAttributeEquals");
+            preferences.remove("tagAttributeValue");
+            preferences.remove("commentBorder");
+            preferences.remove("commentText");
+            preferences.remove("SCRIPT_AREA_BORDER");
+            preferences.remove("jsp_content");
             break;
         }
     }