+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Addded support for the Python editor.
 * Importing themes overwrites stock and previously imported themes.
 * Restoring defaults deletes all imported or replaced themes.
+* Added support for the JSP editor.
 
 0.6.0 (2011-01-26)
 ==================